+++ src/layers/ZweitausbildungAbroadLayer/ZweitausbildungAbroadLayer.js
@@ -38,7 +38,7 @@ class ZweitausbildungAbroadLayer extends VectorLayer {
 
     this.styleCache = {};
 
-    this.setVisible(this.visible);
+    // this.setVisible(this.visible);
   }
 
   setLanguage(language) {